requirements.txt
kafka-python
pandas
pyarrow
pymongo
